modified candidate history module with all data passing through candidate interface and required files are created

+++ candidate_reload_history.py
@@ -3,7 +3,7 @@ import cgi, cgitb
 import sys
 import os
 import MySQLdb
-
+import accept_offer
 def connect_to_database():
 	global conn,cursor
 	conn = MySQLdb.connect (host = "localhost",user = "root",passwd = "",db = "mini_project")
@@ -13,7 +13,7 @@ def connect_to_database():
 def reload_history(c_id1):
 	global cursor,conn
 	connect_to_database()
-	sql="SELECT * FROM History where UserID='%s' and role='Candidate' ORDER BY Time DESC"%(c_id1)
+	sql="SELECT * FROM History where UserID='%s' and role='Candidate' or role='AO' ORDER BY Time DESC"%(c_id1)
 	try:
 		cursor.execute(sql)
 		results = cursor.fetchall()
@@ -22,7 +22,13 @@ def reload_history(c_id1):
 			time=row["Time"]
 			datetime=time.strftime('%H : %M')
 			field=row['Field']
-			fdiv="You have updated <strong> %s</strong> at %s"%(field,datetime)
+			role=row['role']
+			if (role=="AO"):
+				inst_name=accept_offer.get_institute(conn,cursor,field);
+				field_link="""<a id='inst_profile_link' onclick="get_institute_profile(%s)" class='div_link'>%s</a>"""%(field,inst_name)
+				fdiv="You have <strong>acccepted offer from %s</strong> at %s"%(field_link,datetime)
+			else:
+				fdiv="You have updated <strong> %s</strong> at %s"%(field,datetime)
 			print("""<div id="%s"  class="alert alert-info alert-dismissable fade in" ><a href="#" onclick='delete_hist("%s")' class="close" data-dismiss="alert" aria-label="close">&times;</a> %s </div><br/>"""%(divid,divid,fdiv))
 	except:
 		conn.rollback()
@@ -44,7 +50,7 @@ def delete_history(c_id1):
 def delete_all_history(c_id1):
 	global cursor,conn
 	connect_to_database()
-	sql="DELETE FROM History where UserID='%s' and role='Candidate'"%(c_id1)
+	sql="DELETE FROM History where UserID='%s' and role='Candidate' or role='AO'"%(c_id1)
 	try:
 		cursor.execute(sql)
 		conn.commit()
@@ -55,7 +61,7 @@ def delete_all_history(c_id1):
 def history_total_count(c_id1):
 	global cursor,conn
 	connect_to_database()
-	sql="SELECT * FROM History where UserID='%s' and role='Candidate'"%(c_id1)
+	sql="SELECT * FROM History where UserID='%s' and role='Candidate' or role='AO'"%(c_id1)
 	try:
 		cursor.execute(sql)
 		results=cursor.rowcount
